% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_protect.R
\name{check_protect}
\alias{check_protect}
\title{Check that variables in data frame are not restricted.}
\usage{
check_protect(df, remove_protected = FALSE)
}
\arguments{
\item{df}{Data frame to check against set DUA restriction level}

\item{remove_protected}{Will remove protected variables as determined by
DUA restriction level if set to \code{TRUE}. Default behavior is to warn
only.}
}
\description{
Check that variables in data frame are not restricted.
}
