% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify.R
\name{deidentify}
\alias{deidentify}
\title{Convert identifying variable to unique hash}
\usage{
deidentify(df, id_col = NULL, new_id_name = "id", id_length = 64,
  write_crosswalk = FALSE, crosswalk_name = paste0("id_crosswalk_",
  format(Sys.Date(), format = "\%Y\%m\%d")), crosswalk_path = ".")
}
\arguments{
\item{df}{Data frame}

\item{id_col}{String column name with IDs to be replaced. By default is
\code{NULL} and use the value set by the \code{id_column} argument in
\code{set_dua_level()} function.}

\item{new_id_name}{New hashed ID column name}

\item{id_length}{Length of new hashed ID. Cannot be fewer than 12
characters.}

\item{write_crosswalk}{Write crosswalk between old ID and new hash
ID to output window (unless \code{crosswalk_path} is given
value).}

\item{crosswalk_name}{Name of crosswalk file; defaults to generic name
with current date (YYYYMMDD) appended.}

\item{crosswalk_path}{Path to write crosswalk file (CSV). Only used if
\code{write_crosswalk == TRUE}. Defaults to current directory.}
}
\description{
Convert identifying variable to unique hash
}
