% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duawrangler.R
\docType{package}
\name{duawranglr}
\alias{duawranglr}
\alias{duawranglr-package}
\title{duawranglr: Securely Wrangle Dataset According to Data Usage Agreement}
\description{
A set of functions to help users create shareable data sets from
raw data files that contain protected elements. Relying on master
crosswalk files that list restricted variables, package functions
warn users about possible violations of data usage agreement and
prevent writing protected elements.
}
