% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{make_dua_template}
\alias{make_dua_template}
\title{Interactive function to create template file.}
\usage{
make_dua_template(file_name, path = ".", include_notes = TRUE,
  answer_list = NULL)
}
\arguments{
\item{file_name}{Name of template script.}

\item{path}{Path for template script; defaults to current directory.}

\item{include_notes}{If \code{TRUE}, will include notes with notes and
suggestions for filling out the template file.}

\item{answer_list}{List of answer strings to provide if you don't
want to answer questions interactively. Answers must be provided for the
following questions:

1. Do you want to set the DUA crosswalk file? (Yes/No)
1a. DUA crosswalk file (with path):
2. Do the data need to be deidentified? (Yes/No)
2a. Would like to select the ID column now?
2b. ID column name:

If answers to questions (1) and (2) are 'No', then strings for 1a, 2a, and
2b can be left empty since they will be ignored. Default is \code{NULL},
which means questions will be asked interactively.}
}
\description{
\bold{Note} that this function can only be run in interactive mode. Also,
the saved template file is not meant to run as is, but is meant only to
provide a starting structure for the cleaning script.
}
