% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{make_template}
\alias{make_template}
\title{Interactive function to create template file.}
\usage{
make_template(file_name, path = ".", include_notes = TRUE)
}
\arguments{
\item{file_name}{Name of template script.}

\item{path}{Path for template script; defaults to current directory.}

\item{include_notes}{If \code{TRUE}, will include notes with notes and
suggestions for filling out the template file.}
}
\description{
\bold{Note} that this function can only be run in interactive mode. Also,
the saved template file is not meant to run as is, but is meant only to
provide a starting structure for the cleaning script.
}
