% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dua.R
\name{set_dua_cw}
\alias{set_dua_cw}
\title{Function to set data usage agreement data crosswalk}
\usage{
set_dua_cw(dua_cw, path = ".", delimiter = NULL, sheet = NULL,
  ignore_columns = NULL, remap_list = NULL)
}
\arguments{
\item{dua_cw}{Data frame object or file with columns representing
security levels (header equalling name of level) and rows in
each column representing restricted variables}

\item{path}{File path to DUA crosswalk file.}

\item{delimiter}{If reading in a delimited file that is neither a
comma separated value (CSV) nor tab separated value (TSV), a
string value of the delimiter is necessary.}

\item{sheet}{If reading in DUA from Excel file with values not on
the first sheet, supply the sheet number or name}

\item{ignore_columns}{Columns to ignore when reading in DUA crosswalk.}

\item{remap_list}{If raw variable names should be remapped to new names,
provide list with mappings from old names column to new names column:
\code{list('level_one_new' = 'level_one_old')}.}
}
\description{
Function to set data usage agreement data crosswalk
}
