% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dua.R
\name{set_dua_level}
\alias{set_dua_level}
\title{Set data restriction level in system environment variable}
\usage{
set_dua_level(level, deidentify_required = FALSE, id_column = NULL)
}
\arguments{
\item{level}{String value of the data restriction level}

\item{deidentify_required}{Set to \code{TRUE} if ID column must be changed
to protect unique identifier}

\item{id_column}{Column with unique IDs that must be identified if
\code{deidentify_required == TRUE}}
}
\description{
Set data restriction level in system environment variable
}
