% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_dua_df}
\alias{write_dua_df}
\title{Write DUA approved data set}
\usage{
write_dua_df(df, file_name, path = ".", output_type = c("rds", "rdata",
  "csv", "tsv", "delimited", "stata", "sas", "spss"), ...)
}
\arguments{
\item{df}{Data frame object to save.}

\item{file_name}{Name for saved file.}

\item{path}{Path for saved file; default is current directory.}

\item{output_type}{Output data file type. Options include \code{rds}
(DEFAULT), \code{rdata}, \code{csv}, \code{tsv}, \code{delimited},
\code{stata}, and \code{sas}.}

\item{...}{Arguments to pass to write function. Will change depending
on selected \code{output_type}.}
}
\description{
Write DUA approved data set
}
