% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dua_restrictions}
\alias{check_dua_restrictions}
\title{Check data frame columns against currently set restrictions}
\usage{
check_dua_restrictions(df)
}
\arguments{
\item{df}{Data frame to check against set DUA restriction level.}
}
\description{
Once the DUA crosswalk and level have been set, a working data
frame can be checked against active data element restrictions. The
data frame must pass before it can be writen using
\code{write_dua_df()}.
}
\examples{
\donttest{

check_dua_restrictions(df)

}

}
