% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify.R
\name{deid_dua}
\alias{deid_dua}
\title{Convert identifying variable to unique hash}
\usage{
deid_dua(df, id_col = NULL, new_id_name = "id", id_length = 64,
  existing_crosswalk = NULL, write_crosswalk = FALSE,
  crosswalk_name = NULL)
}
\arguments{
\item{df}{Data frame}

\item{id_col}{Column name with IDs to be replaced. By default it is
\code{NULL} and uses the value set by the \code{id_column}
argument in \code{set_dua_level()} function.}

\item{new_id_name}{New hashed ID column name, which must be
different from old name.}

\item{id_length}{Length of new hashed ID; cannot be fewer than 12
characters (default is 64 characters).}

\item{existing_crosswalk}{File name of existing crosswalk. If
existing crosswalk is used, then \code{new_id_name},
\code{id_length}, \code{id_length}, and \code{crosswalk_name}
will be determined by the already existing crosswalk. Arguments
given for these values will be ignored.}

\item{write_crosswalk}{Write crosswalk between old ID and new hash
ID to console (unless \code{crosswalk_path} is given value).}

\item{crosswalk_name}{Name of crosswalk file; defaults to generic
name with current date (YYYYMMDD) appended.}
}
\description{
Convert a column of unique but restricted IDs into a set of new IDs
using secure (SHA-2) hashing algorithm. Users have the option of
saving a crosswalk between the old and new IDs in case observations
need to reidentified at a later date.
}
\examples{
\dontrun{

deid_dua(df, id_col = 'sid', id_length = 20)
deid_dua(df, write_crosswalk = TRUE)
deid_dua(df, existing_crosswalk = './crosswalk/master_crosswalk.csv')

}

}
