% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{dua_write}
\alias{dua_write}
\title{Write securely cleaned data set}
\usage{
dua_write(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Options to pass to base \code{write.table()} function.}
}
\description{
Write securely cleaned data set
}
