% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dua_level.R
\name{see_dua_level}
\alias{see_dua_level}
\title{Show DUA current level setting}
\usage{
see_dua_level(show_restrictions = FALSE, print_width = getOption("width"))
}
\arguments{
\item{show_restrictions}{If \code{TRUE}, show the names of the variables
that are restricted by the current level.}

\item{print_width}{Defaults to global option for screen width.}
}
\description{
Show DUA current level setting
}
