% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dua_level.R
\name{see_dua_options}
\alias{see_dua_options}
\title{Show DUA options}
\usage{
see_dua_options(print_width = getOption("width"))
}
\arguments{
\item{print_width}{Defaults to global option for screen width}
}
\description{
Show DUA options
}
